# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Rasan SaaS Chatbot Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and receive auth cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant details
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDetail"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant basic info
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantRequest"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDetail"
  /v1/tenants/{tenantId}/config:
    put:
      operationId: updateTenantConfig
      tags: [tenants]
      summary: Update tenant bot configuration
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantConfig"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDetail"

  # ─── Documents ────────────────────────────────────────────────────────────────
  /v1/tenants/{tenantId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List knowledge base documents for a tenant
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a knowledge base document
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/tenants/{tenantId}/documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Document deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Conversations ────────────────────────────────────────────────────────────
...
  /super-admin/auth/login:
    post:
      operationId: superAdminLogin
      tags: [super-admin]
      summary: Super admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
        "401":
          $ref: "#/components/responses/Unauthorized"
  /super-admin/tenants:
    get:
      operationId: listAllTenants
      tags: [super-admin]
      summary: List all tenants
      security:
        - cookieAuth: []
      responses:
        "200":
          description: All tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantSummary"

    post:
      operationId: createTenant
      tags: [super-admin]
      summary: Create a new tenant
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantSummary"
  /super-admin/tenants/{tenantId}:
    patch:
      operationId: updateTenantAdmin
      tags: [super-admin]
      summary: Update tenant (admin)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantAdminRequest"
      responses:
        "200":
          description: Updated tenant

    delete:
      operationId: deleteTenant
      tags: [super-admin]
      summary: Delete a tenant
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/tenantId"
      responses:
        "204":
          description: Tenant deleted
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: rasan_token
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
      required: [detail]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        superAdmin:
          type: boolean
...
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
        tenant_id:
          type: string
          format: uuid
        subscription_tier:
          type: string
          enum: [free, starter, pro, business]
      required: [id, email, role]
    TenantConfig:
      type: object
      properties:
        ai_persona_name:
          type: string
        language:
          type: string
          enum: [ar, en, both]
        tone:
          type: string
          enum: [friendly, professional, formal, concise]
        handoff_keywords:
          type: array
          items:
            type: string
        confidence_threshold:
          type: number
          minimum: 0
          maximum: 1
        max_context_turns:
          type: integer
          minimum: 1
          maximum: 30
        telegram_bot_token:
          type: string
    TenantDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        subscription_tier:
          type: string
          enum: [free, starter, pro, business]
        messages_used_this_month:
          type: integer
        is_active:
          type: boolean
        config:
          $ref: "#/components/schemas/TenantConfig"
        created_at:
          type: string
          format: date-time
      required: [id, name, subscription_tier, is_active]
    TenantSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        subscriptionTier:
          type: string
        isActive:
          type: boolean
        messagesUsed:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, subscriptionTier, isActive]
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    UpdateTenantAdminRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subscriptionTier:
          type: string
          enum: [free, starter, pro, business]
        isActive:
          type: boolean
    CreateTenantRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subscriptionTier:
          type: string
          enum: [free, starter, pro, business]
        isActive:
          type: boolean
      required: [name]
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        file_name:
          type: string
        original_name:
          type: string
        file_type:
          type: string
        file_size:
          type: integer
        status:
          type: string
          enum: [queued, processing, ready, error]
        chunk_count:
          type: integer
          nullable: true
        error_message:
          type: string
          nullable: true
        uploaded_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
          nullable: true
      required: [id, file_name, status, uploaded_at]
...